add checks for notes existance

+++ notes/views.py
@@ -8,8 +8,10 @@ def all_notes(req):
     return render(req, 'allnotes.html', {'notes_list': notes_list})
 
 def get_note(req, note_id):
-    note = Note.objects.get(id=note_id)
-    return render(req, 'singlenote.html', {'note': note})
+    if Note.objects.filter(id=note_id).exists():
+        note = Note.objects.get(id=note_id)
+        return render(req, 'singlenote.html', {'note': note})
+    return redirect('/')
 
 def create_note(req):
     if req.method == 'POST':
@@ -20,11 +22,15 @@ def create_note(req):
     return redirect('/')
 
 def remove_note(req, note_id):
-    note = Note.objects.get(id=note_id)
-    note.delete()
+    if Note.objects.filter(id=note_id).exists():
+        note = Note.objects.get(id=note_id)
+        note.delete()
     return redirect('/')
 
 def edit_note(req, note_id):
+    if not Note.objects.filter(id=note_id).exists():
+        return redirect(f'/')
+    
     note = Note.objects.get(id=note_id)
     if req.method == 'POST':
         header = req.POST.get('note-header')
